oh: improve iserdes resets
- add a max_delay constraint to iserdes reset
  they were falling into async_default and were unconstrained,
  leading to ~7ns delays into the reset pin on the gbt reset inputs

- make the reset at least 2 clock cycles long

From the manual:

Every ISERDESE2 in a multiple bit input structure should therefore be
driven by the same reset signal, asserted, and deasserted synchronously
to CLKDIV to ensure that all ISERDESE2 elements come out of reset in
synchronization. The reset signal should only be deasserted when it is
known that CLK and CLKDIV are stable and present, and should be a
minimum of two CLKDIV pulses wide. After deassertion of reset, the
output is not valid until after two CLKDIV cycles.