oh: modify cluster finding oneshot
it now runs only on the fast clock to avoid timing issues thanks to
clock uncertainty.  the output of the oneshot is now a signal only 0.25
bx so it is critical that the latch pulses are timed in correctly. a
constant was added that controls the phase of the latch pulses

additionally, a generic was added that allows this phase to be controlled from
the instantiating module, which will allow the inputs to be not-aligned
to the 40MHz clock