oh: fix the IPBus slave addresses
The IPBus slave addresses are based on IPBus slave indexes, themselves
based on the order in which the "modules nodes" appear in the OptoHybrid
XML address table.

This commit fixes the GBT and MGT module addresses in the XML address
table to match the expectations from the write_package_file.py script
that generates the wb_addr_sel function.

Fixes #31

  The GBT module was configured with addresses decoded by the MGT
  module.

Fixes the parts already implemented in #16.

  The unmatched prefixes are silently ignored, leading to unexpected
  behavior (i.e. successful, but non-working register accesses).