Re-work the GEM loader Virtex-7 part
This is done in an attempt to fix the FIFO that was stuck in write busy
after reset. The long road was taken, replacing the IP FIFO by an XPM
FIFO and making sure the CDC was done properly.

A few fixes and simplifications in the PROM-less code are also included.

The results look quite good with the exception of some issue with the
first AXI burst that can write corrupted data. The valid data follows.
The problem was also present previously and seem to be linked to the
Zynq part of the GEM loader. Will work-around the glitch in software.