Swap the BC0 and BXN[1:0] = 0 frame markers
It was observed that some very noisy trigger links might never lock.
This is due to the comma symbol frame marker having lower precedence
than the overflow frame marker. Therefore no guarantee is given whether
the frame marker is sent or not.

The proposed solution is to use the K.28.5/0xBC comma symbol as the BC0
frame marker. It is always sent periodically on the link (every 3564 BX)
thanks to its highest priority.