Implement phase selection of the asynchronous external trigger
For timing measurements, it is of importance to select the phases in
which the external asynchronous triggers are accepted. It allows to
simulate a synchronous beam.

The implementation proposed here for the CVP13 is rather simple. The
input signal is oversampled 16x using the 320 MHz TTC clock using the
DDR input mode. Is was chosen over the ISERDES for its easier handling,
particularly w.r.t. the timing (that we want constant and
deterministic).

The rising edges are detected in the 40 MHz clock domain. Only the
rising edges in the chosen phases will lead to a trigger signal.