Ensure that failed GBT IC transactions are reported as such
The IPBus slave signals on the register side are defined such that the
write_done signal must be asserted only after the write_pulse signal.
Before this commit, the GBT IC RX implementation was failing to do so
and was presenting the previous transaction status to the write_done
signal. I.e. not sending a pulse, but de-asserting the signal only after
receiving the write_pulse signal. A failed transaction following a
succesfull transaction was not detected.

It was also noted that the the GBT IC output is valid for a single clock
cycle. Make sure to register its output to keep reading out valid data
long after the original transaction.

A side effect of this improvement is the modification of the
EXECUTE_WRITE register behavior. Now, any IC transaction, read or write,
can fail. Please adapt your code in consequence if needed. Note that it
makes error reporting from the firmware to the software much easier to
deal with.

Please also note the register name changes targetting a better
uniformization. The Python scripts have all been adapted.