Use the TTC-based clocks as GBTx/lpGBT common RX clock
In order for the CDC system to be efficient, the common RX clock for the
GBTx and lpGBT cores must be based on the TTC clock. For boards
different than the CTP7, where it was based on the master TX clock, it
was based on the master RX clock. This doesn't allow for a proper CDC
crossing between the read side of the FIFO and the main GEM logic.

Instead, we are here proposing to use straigth the TTC-based clock of
appropriate frequency (120 MHz for the GBTx and 320 MHz for the lpGBT).
The design is simpler and should be a fully working CDC when the RX
sync' FIFO is enabled.