Send last packet from the 10 GbE data formatter
If the event FIFO was empty, the 10 GbE data formatter was not closing
in event in the packet FIFO. The condition to move through the FSM, and
in particular the TRAILER state, was not met (required both the source
FIFO not to be empty and the destination FIFO not to be full). Instead,
consider only the destination FIFO status for the FSM progresses. Note
that such a situation was happening at the end of each run with the last
packet not being sent and release only at the beginning of the
subsequent run.