sbit-mon: avoid overflowing the cluster FIFO
The Xilinx documentation of the XPM_FIFO_SYNC macro reads "A write
operation is performed when the FIFO is not full and wr_en is asserted
on each wr_clk cycle." This behaviour is true regardless of the rd_en
state.

In our specific case, that means that the cluster DAQ FIFO was
systematically overflowing when leaving the FILLING state/entering the
RUNNING state. Would the triggering cluster occur at the same clock
cycle, it would be lost.

Since we must avoid to completely fill the FIFO in the FILLING/RUNNING
states, one must start dropping words out of the FIFO one clock cycle
before the assertion of almost_full. Such a behavior can be implemented
with the help of the prog_full signal. Note the signal assertion
latency: the threshold must be set to SBIT_FIFO_DEPTH - 3 and not
SBIT_FIFO_DEPTH - 2.

With the fix implemented, the frozen clusters now systematically appear
where they should be in the cluster FIFO. More specifically, the readout
window works as follow:

* For a FIFO_TRIGGER_DELAY set to 0 BX, the triggering cluster always
  appears at BX 509.

* For a FIFO_TRIGGER_DELAY set to 509 BX, the triggering cluster always
  appears at BX 0.

* Values are always read in chronological order. Clusters before the
  triggering one are older; clusters after the triggering one are
  younger.

The behavior as well as the data integrity has been checked with an ILA.