Implement a fake multi-BX readout mode
For each L1A, additional consecutive L1A are sent to the whole system,
leading to additional events. The additional L1A are flagged as such in
the event payload in order to help the unpacking software reconstruct
the precise event timing.

While effectively a multi-BX readout, this new mode is qualified of
"fake" since it is incompatible with any DAQLink output. Only the local
DAQ can be used. Additionally, no processing is implemented in firmware
and all decoding and reconstruction has to be done in software.