Check the full VFAT3 RX stream
This commit ensures that all bits sent by a VFAT3 are checekd for
consistency at the receiver. The words must be either part of a DAQ data
packet or a valid VFAT3 DataOut code:

* Idle patterns
* Slow-control codes
* Sync ack or sync verify ack

This is particularly meaningful in the case of the unprotected GE2/1
GBT1, which cannot see its content checked by any other mean.