Ensure 68 working MGT for all CTP7 flavors
Since on the CTP7 the MGT configuration is currently not exposed to the
software stack, this commit makes sure to bump the number of *working*
MGT to 68 for all CTP7 flavors. The new MGT (and assosciated links)
remain unconnected GEx/1. This number of MGT also covers all optical
links and thus should not require a further update.