Attempt to reduce congestion in the CTP7 ME0 flavor
While congestion remains reasonable in the GEx/1 flavors, the router has
more and more troubles routing the ME0 one. Considering the large number
of S-bit and the massive multiplexer used for debugging features, this
is the most likely explanation. Pipeline the multiplexer in multiple
stages.

Also add a generic parameter to disable the raw S-bit testing feature if
deemed needed.