oh: add a CRC to the 4Gbps trigger links
The GE1/1 trigger links are known to be (slightly) instable for multiple
reasons: the infamous initialization issue, VTRx/GBTx problems, SEU,...
On the 4Gbps links, the last 16 bits are unused. This commit adds a CRC
on the last 8 bits to better detect any instability. The end goal is to
discard any corrupted frame and its embedded clusters.

The chosen polynomial is ATM-8/CRC-8P (also known as 0x83 or 0x107),
which is the best for our dataword length following the famous P.
Koopman page. Any error with up to 4 bit flips are detected in the
complete dataword, including the comma word and the unused 8 bits.

See https://users.ece.cmu.edu/~koopman/crc/ for more details about the
choice of the CRC.

Co-authored-by: Andrew Peck <andrew.peck@cern.ch>