lpgbt: register reset signals
This commit aims at relaxing the timing in the critical sections of the
lpgbt module. In particular, everything related to the signals between
the 40 MHz clock domain, usually from an IPBus slave, and the 320 MHz
clock domain, usually in the RX common clock. That has always been a
sore point in the GTY-based builds with those signal often appearing as
the ones with the smallest WNS.

In general, the reset scheme should be reviewed and better registered
since some IPBus registers can trigger an almost complete firmware reset
with very high fanout signals.