gem/daq: fix event building when number of OH is maximum
This is the very first time that we are experimenting with builds
supporting the maximum number of OptoHybrids (i.e. 16 since commit
09c1b7b4). Obviously this is also the moment when we discover bugs in
the corner cases...

In this case, whenever OH #0 was enabled, the event building was stuck
and could not even send a complete event. Whenever OH #0 was disabled,
event building could happen normally.

After debugging, the issue was understood as coming from the lookahead
in state '0x6'. When reaching the last 'e_input_idx' value (i.e. 15),
the lookahead was looking for event availability in chamber #0, wrapping
around. Therefore, a massive, never-ending, event was continuously being
built. After some number of L1A, the AMC13 TTS warning state was
asserted, preventing more data to come in. The DAQ FSM was then stuck in
state '0x4', waiting for events to come from OH #0. In such a situation,
even TTC ReSync were not helpful.

This commit ensures that the lookahead result never looks past the
'e_dav_mask' vector width.