ldaq: fix large events splitting in the 10 GbE data packets formatter
Due to the hard equality, large events could be left not split, leading
to data corruption at the receiver side. More specifically, the (maximum
payload size - 7) had to be a multiple of 4 for the maximum packet size
condition to be met.

This commit fixes the issue by applying an inequality which ensures that
the maximum payload size is never exceeded (by GEM payload data).