gem: workaround corrupted VFAT slow-control EOF
If the VFAT3 slow-control stream is interrupted by a DAQ packet at the
very last bit of the EOF, this last bit is not transmitted. Instead, the
whole EOF is repeated before the next transaction reply.

While the latter situation was already covered, the former was not and
lead to transaction timeouts in presence of L1A.

This commit implements a workaround where the slow-control transaction
can be closed by a DAQ header if all other conditions are already met.