gem: do not trigger automatic actions upon TTC HardReset
[Local reset series 1/3.]

This commit removes all automatic actions taken by the firmware upon TTC
HardReset - which is one of the most general resets possible - from the
GEM flavors. Instead, the recovery mode currently envisioned for GEM is
very local (OptoHybrid, VFAT, OH FPGA,...) and software driven.

In addition to cleaner code and procedures (for example, enabling the
TTC generator only to disable it a few milliseconds later), this change
is hoped to drive developments and testing in the "right direction".

The Python scripts have been adjusted to avoid breaking current tools.