gem: create high-granularity trigger resets
[Local reset series 3/3.]

This commit ensures that all trigger processing modules and counters can
be reset independently. This is typically required to reprogram a single
OptoHybrid in the background of normal data taking.

The changes consist only of additions and should have no functional
impact on the current routines, including the global link and trigger
resets.